+++ core/src/main/scala/org/locationtech/rasterframes/encoders/StandardSerializers.scala
@@ -21,7 +21,6 @@
 
 package org.locationtech.rasterframes.encoders
 
-import org.locationtech.rasterframes.encoders.CatalystSerializer._
 import geotrellis.proj4.CRS
 import geotrellis.raster._
 import geotrellis.spark._
@@ -29,7 +28,8 @@ import geotrellis.spark.tiling.LayoutDefinition
 import geotrellis.vector._
 import org.apache.spark.sql.types._
 import org.locationtech.jts.geom.Envelope
-import org.locationtech.rasterframes.encoders.CatalystSerializer.CatalystIO
+import org.locationtech.rasterframes.TileType
+import org.locationtech.rasterframes.encoders.CatalystSerializer.{CatalystIO, _}
 import org.locationtech.rasterframes.model.LazyCRS
 
 /** Collection of CatalystSerializers for third-party types. */
@@ -244,6 +244,25 @@ trait StandardSerializers {
       io.get[Extent](t, 2),
       io.get[CRS](t, 3),
       io.get[KeyBounds[T]](t, 4)
+    )
+  }
+
+  implicit def rasterSerializer: CatalystSerializer[Raster[Tile]] = new CatalystSerializer[Raster[Tile]] {
+    import org.apache.spark.sql.rf.TileUDT.tileSerializer
+
+    override def schema: StructType = StructType(Seq(
+      StructField("tile", TileType, false),
+      StructField("extent", schemaOf[Extent], false)
+    ))
+
+    override protected def to[R](t: Raster[Tile], io: CatalystIO[R]): R = io.create(
+      io.to(t.tile),
+      io.to(t.extent)
+    )
+
+    override protected def from[R](t: R, io: CatalystIO[R]): Raster[Tile] = Raster(
+      io.get[Tile](t, 0),
+      io.get[Extent](t, 1)
     )
   }