+++ core/src/main/scala/org/locationtech/rasterframes/expressions/package.scala
@@ -28,8 +28,8 @@ import org.apache.spark.sql.catalyst.{InternalRow, ScalaReflection}
 import org.apache.spark.sql.rf.VersionShims._
 import org.apache.spark.sql.{SQLContext, rf}
 import org.locationtech.rasterframes.expressions.accessors._
-import org.locationtech.rasterframes.expressions.aggstats.CellCountAggregate.DataCells
-import org.locationtech.rasterframes.expressions.aggstats._
+import org.locationtech.rasterframes.expressions.aggregates.CellCountAggregate.DataCells
+import org.locationtech.rasterframes.expressions.aggregates._
 import org.locationtech.rasterframes.expressions.generators._
 import org.locationtech.rasterframes.expressions.localops._
 import org.locationtech.rasterframes.expressions.tilestats._
@@ -86,6 +86,7 @@ package object expressions {
     registry.registerExpression[Unequal]("rf_local_unequal")
     registry.registerExpression[Sum]("rf_tile_sum")
     registry.registerExpression[Round]("rf_round")
+    registry.registerExpression[Abs]("rf_abs")
     registry.registerExpression[Log]("rf_log")
     registry.registerExpression[Log10]("rf_log10")
     registry.registerExpression[Log2]("rf_log2")
@@ -120,6 +121,7 @@ package object expressions {
 
     registry.registerExpression[Mask.MaskByDefined]("rf_mask")
     registry.registerExpression[Mask.MaskByValue]("rf_mask_by_value")
+    registry.registerExpression[Mask.InverseMaskByValue]("rf_inverse_mask_by_value")
     registry.registerExpression[Mask.InverseMaskByDefined]("rf_inverse_mask")
 
     registry.registerExpression[DebugRender.RenderAscii]("rf_render_ascii")