+++ pyrasterframes/python/tests/PyRasterFramesTests.py
@@ -191,14 +191,32 @@ class RasterFunctions(TestEnvironment):
             .withColumn('log', rf_log(self.tileCol)) \
             .withColumn('exp', rf_exp(self.tileCol)) \
             .withColumn('expm1', rf_expm1(self.tileCol)) \
-            .withColumn('round', rf_round(self.tileCol))
+            .withColumn('round', rf_round(self.tileCol)) \
+            .withColumn('abs', rf_abs(self.tileCol))
         # TODO: add test for rf_extent and rf_geometry once rastersource connector is integrated and we have
         #  a source of ProjectedRasterTiles.
-        df.show()
+        df.first()
 
     def test_agg_mean(self):
         mean = self.rf.agg(rf_agg_mean(self.tileCol)).first()['rf_agg_mean(tile)']
         self.assertTrue(self.rounded_compare(mean, 10160))
+
+    def test_prt_functions(self):
+        df = self.spark.read.rastersource(self.img_uri) \
+            .withColumn('crs', rf_crs(self.tileCol)) \
+            .withColumn('ext', rf_extent(self.tileCol)) \
+            .withColumn('geom', rf_geometry(self.tileCol))
+        df.show()
+
+    def test_rasterize(self):
+        # NB: This test just makes sure rf_rasterize runs, not that the results are correct.
+        withRaster = self.rf.withColumn('rasterized', rf_rasterize('geometry', 'geometry', lit(42), 10, 10))
+        withRaster.first()
+
+    def test_reproject(self):
+        reprojected = self.rf.withColumn('reprojected', st_reproject('center', 'EPSG:4326', 'EPSG:3857'))
+        reprojected.first()
+
 
     def test_aggregations(self):
         aggs = self.rf.agg(
@@ -282,6 +300,12 @@ class RasterFunctions(TestEnvironment):
         result = rf2.agg(rf_agg_no_data_cells(rf2.tile) < rf_agg_no_data_cells(rf2.masked)) \
             .collect()[0][0]
         self.assertTrue(result)
+
+        rf3 = rf1.select(rf1.tile, rf_inverse_mask_by_value(rf1.tile, rf1.mask, lit(mask_value)).alias('masked'))
+        result = rf3.agg(rf_agg_no_data_cells(rf3.tile) < rf_agg_no_data_cells(rf3.masked)) \
+            .collect()[0][0]
+        self.assertTrue(result)
+
 
     def test_resample(self):
         from pyspark.sql.functions import lit
@@ -546,13 +570,37 @@ class RasterSource(TestEnvironment):
             pathTableColumns=csv_columns,
         )
 
-        # path_df.printSchema()
         self.assertTrue(len(path_df.columns) == 6)  # three bands times {path, tile}
         self.assertTrue(path_df.select('b1_path').distinct().count() == 3)  # as per scene_dict
         b1_paths_maybe = path_df.select('b1_path').distinct().collect()
         b1_paths = [s.format('1') for s in scene_dict.values()]
         self.assertTrue(all([row.b1_path in b1_paths for row in b1_paths_maybe]))
 
+    def test_raster_join(self):
+        # re-read the same source
+        rf_prime = self.spark.read.geotiff(self.img_uri) \
+            .withColumnRenamed('tile', 'tile2').alias('rf_prime')
+
+        rf_joined = self.rf.raster_join(rf_prime)
+
+        self.assertTrue(rf_joined.count(), self.rf.count())
+        self.assertTrue(len(rf_joined.columns) == len(self.rf.columns) + len(rf_prime.columns) - 2)
+
+        rf_joined_2 = self.rf.raster_join(rf_prime, self.rf.extent, self.rf.crs, rf_prime.extent, rf_prime.crs)
+        self.assertTrue(rf_joined_2.count(), self.rf.count())
+        self.assertTrue(len(rf_joined_2.columns) == len(self.rf.columns) + len(rf_prime.columns) - 2)
+
+        # this will bring arbitrary additional data into join; garbage result
+        join_expression = self.rf.extent.xmin == rf_prime.extent.xmin
+        rf_joined_3 = self.rf.raster_join(rf_prime, self.rf.extent, self.rf.crs,
+                                          rf_prime.extent, rf_prime.crs,
+                                          join_expression)
+        self.assertTrue(rf_joined_3.count(), self.rf.count())
+        self.assertTrue(len(rf_joined_3.columns) == len(self.rf.columns) + len(rf_prime.columns) - 2)
+
+        # throws if you don't  pass  in all expected columns
+        with self.assertRaises(AssertionError):
+            self.rf.raster_join(rf_prime, join_exprs=self.rf.extent)
 
 def suite():
     function_tests = unittest.TestSuite()